# add fail condition for invalid perf rawevents and metric events