--- a/perf/perf_metric.py
+++ b/perf/perf_metric.py
@@ -82,12 +82,11 @@ class perf_metric(Test):
 
     def _run_cmd(self, option):
         for line in self.list_of_metric_events:
-            cmd = "perf stat %s %s sleep 1" % (option, line)
-            rc, op = process.getstatusoutput(cmd, ignore_status=True,
-                                             shell=True, verbose=True)
+            cmd = "perf stat %s %s -C 0 sleep 1" % (option, line)
+            op = process.run(cmd, ignore_status=True, shell=True, verbose=True)
+            output = op.stdout.decode() + op.stderr.decode()
             # When the command failed, checking for expected failure or not.
-            if rc:
-                output = op.stdout.decode() + op.stderr.decode()
+            if op.exit_status:
                 found_imc = False
                 found_hv_24_7 = False
                 for ln in output.splitlines():
@@ -107,6 +106,8 @@ class perf_metric(Test):
                                   " environment" % cmd)
                 else:
                     self.fail_cmd.append(cmd)
+            if ("not counted" in output) or ("not supported" in output):
+                self.fail_cmd.append(cmd)
         if self.fail_cmd:
             self.fail("perf_metric: commands failed are %s" % self.fail_cmd)