patch1 adds fail condition for invalid metric events
fix error as process.getstatusoutput has no attribute stdout/stderr
changed perf stat metric events to run with cpu 0

patch2 adds fail condition for invalid perf_rawevents

Signed-off-by: Disha Goel <disgoel@linux.ibm.com>

avocado run --test-runner runner perf_rawevents.py
JOB ID     : 050a3688bba2b2fc14e16aa1532a5f9638480c07
JOB LOG    : /home/disha/avocado-fvt-wrapper/results/job-2023-05-10T18.54-050a368/job.log
 (1/2) perf_rawevents.py:PerfRawevents.test_raw_code: FAIL: perf_raw_events: refer log file for failed events (11.41 s)
 (2/2) perf_rawevents.py:PerfRawevents.test_name_event: PASS (13.63 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/disha/avocado-fvt-wrapper/results/job-2023-05-10T18.54-050a368/results.html
JOB TIME   : 33.78 s
[job.log](https://github.com/avocado-framework-tests/avocado-misc-tests/files/11485888/job.log)

avocado run --test-runner runner perf_metric.py
JOB ID     : 8ca83c528ce357d046902a6cf1ddabf59b79aef9
JOB LOG    : /home/disha/avocado-fvt-wrapper/results/job-2023-05-16T14.47-8ca83c5/job.log
 (1/2) perf_metric.py:perf_metric.test_all_metric_events_with_M: PASS (191.68 s)
 (2/2) perf_metric.py:perf_metric.test_all_metric_events_with_metric: PASS (191.57 s)
RESULTS    : PASS 2 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/disha/avocado-fvt-wrapper/results/job-2023-05-16T14.47-8ca83c5/results.html
JOB TIME   : 391.84 s
[job.log](https://github.com/avocado-framework-tests/avocado-misc-tests/files/11486148/job.log)